# app/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExpenseWise REST API",
    "version": "1.0.0",
    "description": "Developer API for ExpenseWise personal finance tracker. All endpoints return JSON and require authorization headers for private data."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token",
        "description": "Enter your personal secure API Token generated from settings."
      }
    }
  },
  "paths": {
...
    "/v1/expenses": {
      "get": {
        "tags": ["Expenses"],
        "summary": "List and filter user expenses",
        "description": "Retrieves a paginated ledger list of all custom expenses. Decrypts encrypted columns dynamically.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" }, "description": "Page index (1-based)" },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Substring match description/payee" },
          { "name": "category", "in": "query", "schema": { "type": "string" }, "description": "Filter by exact category name" },
          { "name": "start_date", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "YYYY-MM-DD start filter" },
          { "name": "end_date", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "YYYY-MM-DD end filter" }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of expenses.",
            "content": {
              "application/json": {
                "example": {
                  "expenses": [
                    {
                      "id": "e391-4fa1",
                      "amount": 123.45,
                      "category": "Food",
                      "description": "Lunch meeting",
                      "payee": "Subway",
                      "expense_date": "2026-06-15",
                      "payment_mode": "Credit Card"
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "total_pages": 1,
                    "total_items": 1
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Expenses"],
        "summary": "Save a new expense record",
        "description": "Creates and encrypts a new expense entry for the user.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": { "type": "number" },
                  "category": { "type": "string" },
                  "payee": { "type": "string" },
                  "expense_date": { "type": "string", "format": "date" },
                  "description": { "type": "string" },
                  "payment_mode": { "type": "string" }
                },
                "required": ["amount", "category", "expense_date"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense saved successfully.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Expense saved successfully.",
                  "expense": { "id": "e391-4fa1" }
                }
              }
            }
          }
        }
      }
    },
    "/v1/expenses/{uuid}": {
      "get": {
        "tags": ["Expenses"],
        "summary": "Retrieve single expense details",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "uuid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Expense details.",
            "content": {
              "application/json": {
                "example": {
                  "id": "e391-4fa1",
                  "amount": 123.45,
                  "category": "Food",
                  "description": "Lunch meeting",
                  "payee": "Subway",
                  "expense_date": "2026-06-15",
                  "payment_mode": "Credit Card"
                }
              }
            }
          },
          "404": { "description": "Expense not found." }
        }
      },
      "put": {
        "tags": ["Expenses"],
        "summary": "Update an existing expense details",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "uuid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": { "type": "number" },
                  "category": { "type": "string" },
                  "payee": { "type": "string" },
                  "expense_date": { "type": "string", "format": "date" },
                  "description": { "type": "string" },
                  "payment_mode": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense updated successfully.",
            "content": {
              "application/json": {
                "example": { "message": "Expense updated successfully." }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Expenses"],
        "summary": "Delete an expense record",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "uuid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Expense deleted successfully.",
            "content": {
              "application/json": {
                "example": { "message": "Expense deleted successfully." }
              }
            }
          }
        }
      }
    },
...
    "/v1/export": {
      "get": {
        "tags": ["Import/Export"],
        "summary": "Backup the user database to JSON v2.0 file",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Backup payload."
          }
        }
      }
    },
    "/v1/import": {
      "post": {
        "tags": ["Import/Export"],
        "summary": "Restore/merge database from local JSON backup",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Backup restored successfully."
          }
        }
      }
    }
  }
}